Añadida autorizaciones ADMIN

+++ API-Organizador-Eventos/API-Organizador-Eventos/src/main/kotlin/com/example/API_Organizador_Eventos/controller/UsuarioController.kt
@@ -12,10 +12,14 @@ import org.springframework.security.core.Authentication
 import org.springframework.security.core.AuthenticationException
 import org.springframework.web.bind.annotation.*
 import java.util.*
+import org.springframework.security.core.context.SecurityContextHolder
+import org.springframework.security.core.userdetails.UserDetails
+import org.springframework.web.bind.annotation.*
 
 @RestController
 @RequestMapping("/usuarios")
 class UsuarioController {
+
 
     @Autowired
     private lateinit var usuarioService: UsuarioService
@@ -27,12 +31,12 @@ class UsuarioController {
     private lateinit var tokenService: TokenService
 
     @PostMapping("/login")
-    fun login(@RequestBody usuario: Usuario): ResponseEntity<Any>? {
-        println(usuario)
+    fun login(@RequestBody usuario: Usuario): ResponseEntity<Any> {
         val authentication: Authentication
         try {
-            authentication = authenticationManager.authenticate(UsernamePasswordAuthenticationToken(usuario.username, usuario.password))
-            println(authentication.details)
+            authentication = authenticationManager.authenticate(
+                UsernamePasswordAuthenticationToken(usuario.username, usuario.password)
+            )
         } catch (e: AuthenticationException) {
             return ResponseEntity(mapOf("mensaje" to "Credenciales incorrectas"), HttpStatus.UNAUTHORIZED)
         }
@@ -49,13 +53,22 @@ class UsuarioController {
     }
 
     @PostMapping("/signup")
-    fun signup(@RequestBody usuario: Usuario): ResponseEntity<Any>? {
+    fun signup(@RequestBody usuario: Usuario): ResponseEntity<Any> {
         return usuarioService.registrarUsuario(usuario)
     }
 
     @DeleteMapping("/deleteUsuario")
-    fun deleteUsuario(@RequestBody id: Long): ResponseEntity<Any>? {
-        val usuario = usuarioService.deleteUsuario(id)
-        return ResponseEntity(usuario, HttpStatus.OK)
+    fun deleteUsuario(@RequestBody id: Long): ResponseEntity<Any> {
+        val auth = SecurityContextHolder.getContext().authentication
+
+        val usuarioAutenticado = usuarioService.findByUsername(auth.name)
+            ?: return ResponseEntity("Usuario no encontrado", HttpStatus.NOT_FOUND)
+
+        if (usuarioAutenticado.roles=="ADMIN" || usuarioAutenticado.id == id) {
+            val resultado = usuarioService.deleteUsuario(id)
+            return ResponseEntity(resultado, HttpStatus.OK)
+        } else {
+            return ResponseEntity("No tienes permisos para eliminar este usuario", HttpStatus.FORBIDDEN)
+        }
     }
 }
